+++ hazelcast/src/main/java/com/hazelcast/wan/impl/WanNoDelayReplication.java
@@ -106,7 +106,7 @@ public class WanNoDelayReplication
                     Data data = node.nodeEngine.getSerializationService().toData(event);
                     Packet packet = new Packet(data, node.nodeEngine.getPortableContext());
                     packet.setHeader(Packet.HEADER_WAN_REPLICATION);
-                    node.nodeEngine.send(packet, conn);
+                    node.nodeEngine.getPacketTransceiver().transmit(packet, conn);
                 } else {
                     failureQ.addFirst(event);
                     conn = null;
